+++ CHANGELOG.md
@@ -4,3 +4,5 @@
 ## 2021.1
 
 * I created the package.
+
+* I added a parameter for disabling the next/finish button.